Refactor executive summary utilities for Next.js compatibility
